+++ .env.example
@@ -1,0 +1,10 @@
+REDIS_URL=
+
+API_KEY=
+API_SECRET_KEY=
+ACCESS_KEY=
+ACCESS_SECRET=
+
+LOOKUP_TIMEOUT=60
+LOGLEVEL=INFO
+PORT=8000
